requirements.txt
langgraph
langchain-core
langchain-openai
langchain-community
faiss-cpu
pydantic
python-dotenv